Removing video glitches
Glitches occurs on VIDEN and WAIT lines due to the use of asynchronous counters 74HC4040 (the same occurs when using the 74LS393). Those gliteches are usually controlled by a filtering capacitor in the WAIT line. Here a capacitor was also included at the VIDEN line before WAIT signal generation.
Those glitches can be seen as image interference on screen while the CPU are accessing video memory.